% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rand_alphanumeric}
\alias{rand_alphanumeric}
\title{Generate a Random Alphanumeric.}
\usage{
rand_alphanumeric(len = 3)
}
\arguments{
\item{len}{The length of the alphanumeric to produce.}
}
\value{
An alphanumeric string.
}
\description{
Generate a Random Alphanumeric.
}

