% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_dependency_list}
\alias{sbatch_dependency_list}
\title{Fill in dependency list placeholders.}
\usage{
sbatch_dependency_list(dep_list, job_id_sub_history)
}
\arguments{
\item{dep_list}{The string dependency list value of the key-value pair
for a sbatch dependency list options.}

\item{job_id_sub_history}{An array of job ids ordered oldest to newest.}
}
\value{
A dependency list with the proper job ids in the list.
}
\description{
The placeholders \code{PREVIOUS(ALL)} and \code{PREVIOUS(<n>)} can be used in
a sbatch dependency list and are replaced by all or n of the previous
submission job ids.
}

