% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_mail_type_combine}
\alias{sbatch_mail_type_combine}
\title{Create single sbatch mail type key value pair.}
\usage{
sbatch_mail_type_combine(opts)
}
\arguments{
\item{opts}{A list of sbatch mail type options.}
}
\description{
A user may specific multiple \code{sbatch_mail_type_opts},
which must be combined into a single key value pair that
contains the options seperated by commas.
}

