% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\docType{data}
\name{sbatch_mail_type_opts}
\alias{sbatch_mail_type_opts}
\title{A list of sbatch mail types.}
\format{An object of class \code{list} of length 11.}
\usage{
sbatch_mail_type_opts
}
\description{
The value of each item in the list is a string representing
a mail type option.
}
\keyword{datasets}

