% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opt}
\alias{sbatch_opt}
\title{Create an SBATCH option}
\usage{
sbatch_opt(key)
}
\arguments{
\item{key}{The key for the sbatch option.}
}
\value{
A function that takes a single argument representing
the value for the \code{key}.
}
\description{
Create an SBATCH option
}

