% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opt_key}
\alias{sbatch_opt_key}
\title{Get the key of an sbatch.}
\usage{
sbatch_opt_key(opt)
}
\arguments{
\item{opt}{An sbatch option string.}
}
\value{
The \code{opt}'s key.
}
\description{
Get the key of an sbatch.
}

