% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opt_value}
\alias{sbatch_opt_value}
\title{Get the value of an sbatch option.}
\usage{
sbatch_opt_value(opt)
}
\arguments{
\item{opt}{An sbatch option string.}
}
\value{
The \code{opt}'s value.
}
\description{
Get the value of an sbatch option.
}

