% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\docType{data}
\name{sbatch_opts}
\alias{sbatch_opts}
\title{A list of sbatch options.}
\format{An object of class \code{list} of length 8.}
\usage{
sbatch_opts
}
\description{
The value of each item in the list is a string or a function
which takes a string as a parameter, using \code{sbatch_opt}.
}
\details{
To ensure proper formatting, the \code{mail_type} option should
be set using \code{sbatch_mail_types}. Multiple mail types need
to be comma seperated.
}
\keyword{datasets}

