% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opts_equal}
\alias{sbatch_opts_equal}
\title{Test sbatch options for equality.}
\usage{
sbatch_opts_equal(opt_1, opt_2)
}
\arguments{
\item{opt_1}{An sbatch option string.}

\item{opt_2}{An sbatch option string.}
}
\value{
A boolean value.
}
\description{
sbatch option equallity is achieved if the keys of the options
are the same.
}

