% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opts_format_cmd_line}
\alias{sbatch_opts_format_cmd_line}
\title{Format the sbatch options for command line use.}
\usage{
sbatch_opts_format_cmd_line(opts)
}
\arguments{
\item{sbatch_options}{A data table with two columns, \code{option} and
\code{param}.}
}
\description{
Format the sbatch options for command line use.
}

