% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbatch.r
\name{sbatch_opts_insert}
\alias{sbatch_opts_insert}
\title{Insert an sbatch option into a set.}
\usage{
sbatch_opts_insert(opt, opts = c())
}
\arguments{
\item{opt}{The sbatch option to insert}

\item{opts}{A set of sbatch options. Default value is the empty
set.}
}
\value{
A set with \code{opt} inserted.
}
\description{
Insert an sbatch option into a set.
}

