% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slog.r
\name{slog}
\alias{slog}
\title{Slurm Logging}
\usage{
slog(..., slog_file = .__current_stain_log__)
}
\arguments{
\item{...}{Zero or more objects which can be coerced to character.}

\item{slog_file}{A path to a log file Defaults to
\code{.__current_stain_log__}.}
}
\description{
A log can be kept for a certain slurm job. A Stain object must be provided or
else the \code{.__current_stain_object__} will be used
}

