% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slog.r
\name{slog_file}
\alias{slog_file}
\title{Slurm Logging File}
\usage{
slog_file(stain, log_id)
}
\arguments{
\item{stain}{A Stain object.}

\item{log_id}{A unique identifier for the log file.}
}
\description{
Slurm Logging File
}

