% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-bash-script.r
\docType{data}
\name{SlurmBashScript}
\alias{SlurmBashScript}
\title{SlurmBashScript R6 object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SlurmBashScript
}
\description{
Generates the necessary bash script to submit through
the `sbatch` command.
}
\keyword{datasets}

