% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-job.r
\docType{data}
\name{SlurmJob}
\alias{SlurmJob}
\title{SlurmJob R6 object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SlurmJob
}
\description{
An interface to SLURM bash scripts and their submissions.
}
\keyword{datasets}

