% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-options.r
\docType{data}
\name{SlurmOptions}
\alias{SlurmOptions}
\title{SlurmOptions R6 object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SlurmOptions
}
\description{
An interface to SBATCH settings.
}
\keyword{datasets}

