% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-settings.r
\docType{data}
\name{SlurmSettings}
\alias{SlurmSettings}
\title{SlurmSettings R6 object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SlurmSettings
}
\description{
An interface to SBATCH settings.
}
\keyword{datasets}

