% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-container.r
\docType{data}
\name{Stain}
\alias{Stain}
\title{SlurmContainer R6 object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Stain
}
\description{
A slurm container is simply a directory with a specific
structure, particulary it has a submit.slurm script at the
top level.
}
\keyword{datasets}

