% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-bash-script.r
\name{stain_default_main_write}
\alias{stain_default_main_write}
\title{Write the default main R file.}
\usage{
stain_default_main_write(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\description{
Write the default main R file.
}

