% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-message.r
\name{stain_message_globals}
\alias{stain_message_globals}
\title{Message for globals.}
\usage{
stain_message_globals(globals)
}
\arguments{
\item{globals}{The list of globals for a stain.}
}
\description{
Prompt the user to set globals if they have not already been
specified.
}

