% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-message.r
\name{stain_message_source_files}
\alias{stain_message_source_files}
\title{Message for source files.}
\usage{
stain_message_source_files(source_files)
}
\arguments{
\item{source_files}{The list of R source files.}
}
\description{
One of the source files must contain a \code{main} function and this
message will notify the user if none of his or her source files
contain a \code{main} function.
}

