% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_create}
\alias{stain_meta_create}
\title{Create the meta JSON file.}
\usage{
stain_meta_create(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\description{
Create the meta JSON file.
}

