% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_read}
\alias{stain_meta_read}
\title{Read tne stain meta file.}
\usage{
stain_meta_read(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\value{
An R object representing the JSON meta data.
}
\description{
Read tne stain meta file.
}

