% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_sbatch_opts_read}
\alias{stain_meta_sbatch_opts_read}
\title{Get the sbatch options in meta file.}
\usage{
stain_meta_sbatch_opts_read(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\value{
A collection of sbatch options formatted by \code{sbatch_opt}.
}
\description{
Get the sbatch options in meta file.
}

