% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_set_id}
\alias{stain_meta_set_id}
\title{Set the id for the stain.}
\usage{
stain_meta_set_id(dir, id)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}

\item{id}{A 64 character alphanumeric.}
}
\description{
Set the id for the stain.
}

