% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_set_sbatch_opt}
\alias{stain_meta_set_sbatch_opt}
\title{Set a sbatch option parameter.}
\usage{
stain_meta_set_sbatch_opt(dir, opt, param)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}

\item{opt}{The sbatch option key.}

\item{param}{The value for the sbatch option.}
}
\value{
A single string of sbatch options ready for the command line.
}
\description{
Set a sbatch option parameter.
}

