% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_set_stain_version}
\alias{stain_meta_set_stain_version}
\title{Set the stain version.}
\usage{
stain_meta_set_stain_version(dir, stain_version)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}

\item{stain_version}{The version of the Stain package.}
}
\description{
Set the stain version.
}

