% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_sub_history_append}
\alias{stain_meta_sub_history_append}
\title{Append a new submission history row.}
\usage{
stain_meta_sub_history_append(dir, job_id,
  sub_date = as.character(Sys.time()))
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}

\item{job_id}{The slurm job identifier.}

\item{sub_date}{The submission date for the corresponding slurm job. Default
value is \code{Sys.time()}.}
}
\description{
Append a new submission history row.
}

