% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-meta.r
\name{stain_meta_write}
\alias{stain_meta_write}
\title{Write tne stain meta file.}
\usage{
stain_meta_write(meta, dir)
}
\arguments{
\item{meta}{The meta data R object.}

\item{dir}{The location of the \code{.stain/} directory.}
}
\description{
This method should not be used directly by functions outside this file.
}

