% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain.r
\name{stain_new}
\alias{stain_new}
\title{Create a new slurm container.}
\usage{
stain_new(location = ".", source_files = list(),
  settings = SlurmSettings$new())
}
\arguments{
\item{location}{The directory where to create the slurm container. The
default location is the current working directory.}

\item{source_files}{R source files.}

\item{settings}{A \code{SlurmSettings} object.}
}
\value{
A \code{SlurmContainer} object.
}
\description{
Create a new slurm container.
}

