% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_scp}
\alias{stain_scp}
\title{scp with the Stain RSA key.}
\usage{
stain_scp(from, to)
}
\arguments{
\item{from}{The directory or file to copy.}

\item{to}{The destination.}
}
\description{
The stain-specific key must be used to ensure remote login.
}

