% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_ssh}
\alias{stain_ssh}
\title{ssh with the Stain RSA key.}
\usage{
stain_ssh(user, host, cmds = "", intern = FALSE)
}
\arguments{
\item{user}{The user on your remote host.}

\item{host}{The static ip address or url for the remote host.}

\item{cmds}{A sting of one or more commands to run on the remote host.}

\item{intern}{Indicates whether to capture the output of the command
as an R character vector.}
}
\description{
The stain-specific key must be used to ensure remote login.
}

