% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_ssh_key_exists}
\alias{stain_ssh_key_exists}
\title{Check for a Stain ssh key.}
\usage{
stain_ssh_key_exists()
}
\value{
If a public/private key pair exists in \code{~/.ssh/} with the name
\code{stain_rsa}, return TRUE, otherwise return FALSE.
}
\description{
Check for a Stain ssh key.
}

