% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_ssh_key_gen}
\alias{stain_ssh_key_gen}
\title{Generate a Stain ssh key.}
\usage{
stain_ssh_key_gen(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Should an existing Stain ssh key be overwritten. Default
value is FALSE.}
}
\description{
A 4096 bit key will be generated and stored in \code{~/.ssh/} with the name
\code{stain_rsa}.
}

