% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_ssh_setup}
\alias{stain_ssh_setup}
\title{Create bash code for ssh setup.}
\usage{
stain_ssh_setup(user, host)
}
\arguments{
\item{user}{The user on your remote host.}

\item{host}{The static ip address or url for the remote host.}
}
\value{
A single bash command to run.
}
\description{
In order for a remote submission to work, an ssh public key for Stain must
be present in the remote host's \code{~/.ssh/authorized_keys} list. This
process requires two steps. 1) To \code{scp} the public key and 2) to add
the key to \code{~/.ssh/authorized_keys}. This function will autogenerate
the necessary bash code to complete these steps.
}

