% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_ssh_squeue}
\alias{stain_ssh_squeue}
\title{Get squeue info on certain jobs.}
\usage{
stain_ssh_squeue(user, host, job_ids)
}
\arguments{
\item{job_ids}{A collection of job ids for which to fetch statuses.}
}
\value{
A data frame with columns corresponding to those produced
by the \code{squeue -l} command.
}
\description{
Get squeue info on certain jobs.
}

