% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-history.r
\name{stain_sub_history}
\alias{stain_sub_history}
\title{Stain's submission history.}
\usage{
stain_sub_history(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\value{
A data frame with \code{job_id} and \code{submission_date} columns,
both of type "character".
}
\description{
Stain's submission history.
}

