% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-history.r
\name{stain_sub_history_append}
\alias{stain_sub_history_append}
\title{Append a job id to submission history.}
\usage{
stain_sub_history_append(dir, job_id, sub_date = Sys.time())
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}

\item{job_id}{The slurm job identifier.}

\item{sub_date}{The timestamp when the slurm job for the corresponding
\code{job_id} was submitted. Defaults to the system's current time.}
}
\description{
Append a job id to submission history.
}

