% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm-job.r
\name{submit_jobs}
\alias{submit_jobs}
\title{Submit one or more slurm jobs.}
\usage{
submit_jobs(jobs)
}
\arguments{
\item{jobs}{The \code{job_<alphanumeric>/} directories for
the slurm container. May also}
}
\description{
This function will submit your slurm job given the path
to a slurm container.
}

