% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slog.r
\name{slog}
\alias{slog}
\title{Slurm Logging}
\usage{
slog(..., stain = .__current_stain_object__)
}
\arguments{
\item{...}{Zero or more objects which can be coerced to character.}

\item{stain}{A Stain object. Defaults to \code{.__current_stain_object__}.}
}
\description{
A log can be kept for a certain slurm job. A Stain object must be provided or
else the \code{.__current_stain_object__} will be used
}

