% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-script-writers.r
\name{stain_bash_slurm_write}
\alias{stain_bash_slurm_write}
\title{Write the stain slurm submission script}
\usage{
stain_bash_slurm_write(dir)
}
\arguments{
\item{dir}{The location of the \code{.stain/} directory.}
}
\description{
Write the stain slurm submission script
}

