% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stain-ssh.r
\name{stain_scp}
\alias{stain_scp}
\title{scp with the Stain RSA key.}
\usage{
stain_scp(user, host, from, to)
}
\arguments{
\item{user}{The user on your remote host.}

\item{host}{The static ip address or url for the remote host.}

\item{from}{The local files or directory to copy from.}

\item{to}{The directory in \code{<user>@<host>} to copy into.}
}
\description{
The stain-specific key must be used to ensure remote login.
}

